Ghassen

+++ AngularFront/src/app/AccueilClient/card-cart/card-cart.component.ts
@@ -13,6 +13,7 @@ export class CardCartComponent implements OnInit{
 @Input() products:any
 length:any
 total:number=0
+
 updateQte(product: any) {
   this.productServiceStorage.updateQte(product);
   location.reload();
@@ -22,15 +23,26 @@ refresh(){
 }
 
 deleteProduct(product:any){
-  this.productServiceStorage.deleteProduct(product)
+  if(this.productServiceStorage.GetProduct()?.length>1){
+    this.productServiceStorage.deleteProduct(product)
+  }else{
+    this.productServiceStorage.clearProductList();
+  }
   this.location.replaceState('/cart');
   location.reload();
 }
+plusQte(product:any){
+  product.quantity +=1;
+}
+
+moinsQte(product:any){
+  if(product.quantity>0)
+  product.quantity -=1;
+}
 
 ngOnInit(): void {
-  this.length=this.products.length;
-  console.log(this.products);
-  for (let index = 0; index < this.products.length; index++) {
+  this.length=this.products?.length;
+  for (let index = 0; index < this.products?.length; index++) {
     this.total = Number(this.products[index].prix) * Number(this.products[index].quantity)
   }
 }
